Langchain_One_Shot/requirements.txt
langchain
openai
huggingface_hub
streamlit
python-dotenv


